requirements.txt
fastapi
uvicorn[standard]
pydantic
pandas
python-docx
pymongo
numpy
scikit-learn
sentence-transformers
python-dotenv
langchain
openai
ollama
langchain==0.2.15
langchain-core==0.2.43
langchain-community==0.2.15
langchain-openai==0.1.25
langchain-ollama==0.1.3
openpyxl>=3.1.3
langchain-mongodb